[SPOT-280][REFACTOR] Organize Study Domain
