+++ src/main/java/com/example/spot/legacy/service/notification/NotificationCommandServiceImpl.java
@@ -3,10 +3,10 @@ package com.example.spot.legacy.service.notification;
 import com.example.spot.refactor.common.api.code.status.ErrorStatus;
 import com.example.spot.refactor.common.api.exception.GeneralException;
 import com.example.spot.legacy.domain.Notification;
-import com.example.spot.refactor.study.domain.aggregate.studymember.StudyMember;
+import com.example.spot.refactor.study.domain.aggregate.StudyMember;
 import com.example.spot.refactor.study.domain.enums.StudyApplicationStatus;
 import com.example.spot.legacy.domain.enums.NotifyType;
-import com.example.spot.refactor.study.domain.aggregate.studymember.StudyMemberRepository;
+import com.example.spot.refactor.study.domain.aggregate.StudyMemberRepository;
 import com.example.spot.legacy.web.dto.notification.NotificationResponseDTO.NotificationProcessDTO;
 import java.util.Objects;
 import lombok.RequiredArgsConstructor;
