+++ src/main/java/com/example/spot/refactor/member/domain/association/MemberTheme.java
@@ -1,7 +1,7 @@
 package com.example.spot.refactor.member.domain.association;
 
 import com.example.spot.refactor.member.domain.Member;
-import com.example.spot.refactor.study.domain.aggregate.studytheme.Theme;
+import com.example.spot.refactor.study.domain.aggregate.Theme;
 import com.example.spot.refactor.common.entity.BaseEntity;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
