+++ src/main/java/com/example/spot/refactor/member/domain/association/PreferredRegion.java
@@ -1,6 +1,6 @@
 package com.example.spot.refactor.member.domain.association;
 
-import com.example.spot.refactor.study.domain.aggregate.studyregion.Region;
+import com.example.spot.refactor.study.domain.aggregate.Region;
 import com.example.spot.refactor.member.domain.Member;
 import com.example.spot.refactor.common.entity.BaseEntity;
 import jakarta.persistence.*;
