+++ src/main/java/com/example/spot/refactor/study/domain/aggregate/StudyTheme.java
@@ -1,7 +1,6 @@
-package com.example.spot.refactor.study.domain.aggregate.studytheme;
+package com.example.spot.refactor.study.domain.aggregate;
 
 import com.example.spot.refactor.common.entity.BaseEntity;
-import com.example.spot.refactor.study.domain.aggregate.Study;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
