+++ src/main/java/com/example/spot/refactor/study/domain/repository/StudyRepositoryCustom.java
@@ -1,10 +1,10 @@
 package com.example.spot.refactor.study.domain.repository;
 
 import com.example.spot.refactor.member.domain.enums.Status;
-import com.example.spot.refactor.study.domain.aggregate.studymember.StudyMember;
+import com.example.spot.refactor.study.domain.aggregate.StudyMember;
 import com.example.spot.refactor.study.domain.enums.StudySortBy;
-import com.example.spot.refactor.study.domain.aggregate.studyregion.StudyRegion;
-import com.example.spot.refactor.study.domain.aggregate.studytheme.StudyTheme;
+import com.example.spot.refactor.study.domain.aggregate.StudyRegion;
+import com.example.spot.refactor.study.domain.aggregate.StudyTheme;
 import com.example.spot.refactor.study.domain.aggregate.Study;
 import java.util.List;
 import java.util.Map;
