+++ src/test/java/com/example/spot/service/notification/NotificationCommandServiceTest.java
@@ -11,9 +11,9 @@ import com.example.spot.refactor.member.domain.Member;
 import com.example.spot.legacy.domain.Notification;
 import com.example.spot.refactor.study.domain.enums.StudyApplicationStatus;
 import com.example.spot.legacy.domain.enums.NotifyType;
-import com.example.spot.refactor.study.domain.aggregate.studymember.StudyMember;
+import com.example.spot.refactor.study.domain.aggregate.StudyMember;
 import com.example.spot.refactor.study.domain.aggregate.Study;
-import com.example.spot.refactor.study.domain.aggregate.studymember.StudyMemberRepository;
+import com.example.spot.refactor.study.domain.aggregate.StudyMemberRepository;
 import com.example.spot.legacy.repository.NotificationRepository;
 import com.example.spot.legacy.service.notification.NotificationCommandServiceImpl;
 import com.example.spot.legacy.web.dto.notification.NotificationResponseDTO.NotificationProcessDTO;
