+++ src/test/java/com/example/spot/service/study/StudyCommandServiceTest.java
@@ -4,25 +4,25 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.Mockito.*;
 
-import com.example.spot.refactor.study.domain.aggregate.studyregion.StudyRegion;
+import com.example.spot.refactor.study.domain.aggregate.StudyRegion;
 import com.example.spot.refactor.common.api.exception.handler.StudyHandler;
 import com.example.spot.refactor.member.domain.Member;
-import com.example.spot.refactor.study.domain.aggregate.studyregion.Region;
-import com.example.spot.refactor.study.domain.aggregate.studymember.StudyMember;
-import com.example.spot.refactor.study.domain.aggregate.studytheme.Theme;
+import com.example.spot.refactor.study.domain.aggregate.Region;
+import com.example.spot.refactor.study.domain.aggregate.StudyMember;
+import com.example.spot.refactor.study.domain.aggregate.Theme;
 import com.example.spot.refactor.study.domain.enums.StudyApplicationStatus;
 import com.example.spot.refactor.member.domain.enums.Gender;
 import com.example.spot.refactor.study.domain.enums.StudyState;
 import com.example.spot.refactor.study.domain.enums.ThemeType;
-import com.example.spot.refactor.study.domain.aggregate.studytheme.StudyTheme;
+import com.example.spot.refactor.study.domain.aggregate.StudyTheme;
 import com.example.spot.refactor.study.domain.aggregate.Study;
 import com.example.spot.refactor.member.domain.MemberRepository;
-import com.example.spot.refactor.study.domain.aggregate.studymember.StudyMemberRepository;
-import com.example.spot.refactor.study.domain.aggregate.studyregion.RegionRepository;
-import com.example.spot.refactor.study.domain.aggregate.studyregion.StudyRegionRepository;
+import com.example.spot.refactor.study.domain.aggregate.StudyMemberRepository;
+import com.example.spot.refactor.study.domain.aggregate.RegionRepository;
+import com.example.spot.refactor.study.domain.aggregate.StudyRegionRepository;
 import com.example.spot.refactor.study.domain.repository.StudyRepository;
-import com.example.spot.refactor.study.domain.aggregate.studytheme.StudyThemeRepository;
-import com.example.spot.refactor.study.domain.aggregate.studytheme.ThemeRepository;
+import com.example.spot.refactor.study.domain.aggregate.StudyThemeRepository;
+import com.example.spot.refactor.study.domain.aggregate.ThemeRepository;
 import com.example.spot.refactor.study.application.StudyCommandServiceImpl;
 import com.example.spot.refactor.study.presentation.dto.request.StudyJoinRequestDTO;
 import com.example.spot.refactor.study.presentation.dto.request.StudyRegisterRequestDTO;
