+++ src/test/java/com/example/spot/service/study/studymember/StudyMemberQueryServiceTest.java
@@ -2,7 +2,7 @@ package com.example.spot.service.study.studymember;
 
 import com.example.spot.refactor.common.api.exception.GeneralException;
 import com.example.spot.refactor.member.domain.Member;
-import com.example.spot.refactor.study.domain.aggregate.studymember.StudyMember;
+import com.example.spot.refactor.study.domain.aggregate.StudyMember;
 import com.example.spot.refactor.study.domain.aggregate.studyschedule.StudySchedule;
 import com.example.spot.refactor.study.domain.aggregate.studytodo.StudyToDo;
 import com.example.spot.refactor.study.domain.enums.StudyApplicationStatus;
@@ -10,7 +10,7 @@ import com.example.spot.refactor.study.domain.enums.StudyPostCategory;
 import com.example.spot.refactor.study.domain.aggregate.Study;
 import com.example.spot.refactor.study.domain.aggregate.studypost.StudyPost;
 import com.example.spot.refactor.member.domain.MemberRepository;
-import com.example.spot.refactor.study.domain.aggregate.studymember.StudyMemberRepository;
+import com.example.spot.refactor.study.domain.aggregate.StudyMemberRepository;
 import com.example.spot.refactor.study.domain.aggregate.studyschedule.StudyScheduleRepository;
 import com.example.spot.refactor.study.domain.aggregate.studypost.StudyPostRepository;
 import com.example.spot.refactor.study.domain.aggregate.studytodo.StudyToDoRepository;
