+++ src/test/java/com/example/spot/service/study/studypost/StudyPostQueryServiceTest.java
@@ -2,7 +2,7 @@ package com.example.spot.service.study.studypost;
 
 import com.example.spot.refactor.common.api.exception.handler.StudyHandler;
 import com.example.spot.refactor.member.domain.Member;
-import com.example.spot.refactor.study.domain.aggregate.studymember.StudyMember;
+import com.example.spot.refactor.study.domain.aggregate.StudyMember;
 import com.example.spot.refactor.study.domain.aggregate.studypost.*;
 import com.example.spot.refactor.study.domain.enums.StudyApplicationStatus;
 import com.example.spot.refactor.member.domain.enums.Gender;
@@ -11,7 +11,7 @@ import com.example.spot.refactor.study.domain.enums.StudyPostCategoryQuery;
 import com.example.spot.refactor.study.domain.aggregate.studypost.LikedStudyComment;
 import com.example.spot.refactor.study.domain.aggregate.Study;
 import com.example.spot.refactor.member.domain.MemberRepository;
-import com.example.spot.refactor.study.domain.aggregate.studymember.StudyMemberRepository;
+import com.example.spot.refactor.study.domain.aggregate.StudyMemberRepository;
 import com.example.spot.refactor.study.domain.repository.StudyRepository;
 import com.example.spot.refactor.study.application.StudyPostQueryServiceImpl;
 import com.example.spot.refactor.study.presentation.dto.response.StudyPostCommentResponseDTO;
